rename rooms -> game_rooms
