+++ app/actions/randomize_teams_action.py
@@ -5,7 +5,7 @@ from attr import evolve
 
 from app.actions.validation_utils import validate_fields
 from app.model import fields
-from app.model.rooms import game_room_exists, get_room_state, update_room
+from app.model.game_rooms import game_room_exists, get_room_state, update_room
 
 
 def randomize_teams(room_name: str) -> Dict[str, str]:
