+++ app/actions/unready_phrases_action.py
@@ -1,7 +1,7 @@
 from typing import Dict, Any
 
 from app.model.player import build_player
-from app.model.rooms import get_room_state
+from app.model.game_rooms import get_room_state
 
 
 def unready_phrases(room_name, player_name) -> Dict[str, Any]:
