+++ app/test_game.py
@@ -3,8 +3,8 @@ from attr import evolve
 
 from app.model import fields
 from app.model.player import build_player
-from app.model.rooms import (game_room_exists, get_room_state,
-                             initialize_game_room, update_room)
+from app.model.game_rooms import (game_room_exists, get_room_state,
+                                  initialize_game_room, update_room)
 
 
 def create_test_game() -> None:
